Extract the serde functionality to a separate module.
